refactor(env-prepare): upgrade install `7-zip` scripts

+++ env-prepare/src/git-install-2.ps1
@@ -1,14 +1,19 @@
-. "./git-install-common.ps1"
+. "$PSScriptRoot\.utils\index.ps1"
+
+. "$PSScriptRoot\git-install-base.ps1"
 
 $unpackgeDir="$downloadPath\PortableGit"
-$installDir="$env:PROGRAMFILES\Git\PortableGit-$version-64-bit"
 
-if (Test-Path   -Path "$installDir" -PathType Container) {
+if ($null -eq $WIN_PROGRAM_FILES) {
+  Write-Host "Not in Windows, skip."
+} else {
+  $installDir="$WIN_PROGRAM_FILES\Git\PortableGit-$version-64-bit"
+  if (Test-Path   -Path "$installDir" -PathType Container) {
     Remove-Item -Path "$installDir" -Recurse -Force
+  }
+  cp -r  $unpackgeDir $installDir
 }
 
-cp -r  $unpackgeDir $installDir
+setMachineEnvPath 'GIT_PATH' "$installDir\bin"
 
-[Environment]::SetEnvironmentVariable('GIT_PATH', "$installDir\bin", 'Machine')
-
-echo "Please add 'GIT_PATH' to 'PATH'"
+Write-Host "Please add 'GIT_PATH' to 'PATH'"
